chore(config): rename validate module to resolve
